+++ CHANGELOG.md
@@ -9,6 +9,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 ### Added
+- Add support for Pattern Artifacts, which allows one Artifact definition to match many actual
+  artifacts; see the documentation of `cache::Artifacts` for details.
 
 ### Changed
 - `cache::Cache`: `artifacts` field is now private.  Get an individual artifact by name using the